+++ admin/group_delete.php
@@ -1,10 +1,16 @@
 <?php
 require_once "../connection.php";
+require_once "../classes/classes.php";
 $id = $_GET['id'];
-$sql = "DELETE FROM groups WHERE id = $id";
-$query = mysqli_query($conn, $sql);
-if (!$query) {
-  echo "Error: " . $sql . "<br>" . mysqli_error($conn);
-} else {
-  header("Location: groups.php");
+
+$group = new Groups;
+
+if ($group->ifId($id)) {
+  $sql = "DELETE FROM groups WHERE id = $id";
+  $query = mysqli_query($conn, $sql);
+  if (!$query) {
+    echo "Error: " . $sql . "<br>" . mysqli_error($conn);
+  } else {
+    header("Location: groups.php");
+  }
 }
